requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
langchain
langchain-openai
chromadb
pypdf
unstructured[pdf]
langchain-community
python-multipart
langchain-chroma
langchain-huggingface
sentence-transformers
langfuse